update:#16 メールアドレス検証、入力がからかを検証するメソッドをoperationに移動 サインイン、サインアウトで共通で利用するため移動
