+++ src/components/templates/SignUp/hook.ts
@@ -1,29 +1,7 @@
-/**
- * メールアドレスの検証
- * @param email - 検証したいEメール
- * @return (true: OK, false: NG)
- */
-export const isValidEmailFormat = (email: string) => {
-  const regex =
-    /^[a-zA-Z0-9_.+-]+@([a-zA-Z0-9][a-zA-Z0-9-]*[a-zA-Z0-9]*\.)+[a-zA-Z]{2,}$/;
-  return regex.test(email);
-};
-
-/**
- * 入力値の配列に空文字があるかを判断する、
- * 一つでも空文字があれば、falseを返す
- * @param args - 文字列の配列
- * @return (true: 空文字がない, false: 空文字がある)
- */
-export const isValidRequiredInput = (args: string[]): boolean => {
-  let validator = true;
-  args.forEach((arg: string) => {
-    if (arg === '') {
-      validator = false;
-    }
-  });
-  return validator;
-};
+import {
+  isValidRequiredInput,
+  isValidEmailFormat,
+} from '../../../reducks/users/operation';
 
 /**
  * 入力データが正しいフォーマットか検証
